+++ src/main/java/org/bytedeco/javacpp/tools/Info.java
@@ -53,6 +53,8 @@ public class Info {
         define = i.define;
         enumerate = i.enumerate;
         flatten = i.flatten;
+        immutable = i.immutable;
+        beanify = i.beanify;
         objectify = i.objectify;
         translate = i.translate;
         skip = i.skip;
@@ -97,6 +99,8 @@ public class Info {
     boolean flatten = false;
     /** Disables generation of setters for public data members of a class */
     boolean immutable = false;
+    /** Adds JavaBeans-style prefixes to getters and setters of public data members of a class */
+    boolean beanify = false;
     /** Map global functions to instance methods, without {@code static} modifier, to implement an interface, etc. */
     boolean objectify = false;
     /** Attempts to translate naively the statements of variable-like macros to Java. */
@@ -134,6 +138,8 @@ public class Info {
     public Info flatten(boolean flatten) { this.flatten = flatten; return this; }
     public Info immutable() { this.immutable = true; return this; }
     public Info immutable(boolean immutable) { this.immutable = immutable; return this; }
+    public Info beanify() { this.beanify = true; return this; }
+    public Info beanify(boolean beanify) { this.beanify = beanify; return this; }
     public Info objectify() { this.objectify = true; return this; }
     public Info objectify(boolean objectify) { this.objectify = objectify; return this; }
     public Info translate() { this.translate = true; return this; }